requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
black==22.3.0
cffi==1.15.0
click==8.1.3
fastapi==0.78.0
gevent==21.12.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websocket==0.2.1
websockets==10.3
zope.event==4.5.0
zope.interface==5.4.0
